fix: fix ordering of logs

+++ api/core/dependencies/job_runner/tasks/ffmpeg_tools/audio_extractor.py
@@ -66,4 +66,5 @@ except Exception as e:
     raise e
 
 finally:
-    delete_file(video_file)
+    if video_file:
+        delete_file(video_file)
